Fix typings in db

+++ backend/src/models/traffic.py
@@ -142,19 +142,19 @@ class service_performance(SQLModel, table=True):
     scenario: str = Field(primary_key=True)
     year: int = Field(primary_key=True)
     kmuoto: str = Field(primary_key=True)
-    myymal_hyper: int
-    myymal_super: int
-    myymal_pien: int
-    myymal_muu: int
-    majoit: int
-    asla: int
-    ravint: int
-    tsto: int
-    liiken: int
-    hoito: int
-    kokoon: int
-    opetus: int
-    muut: int
+    myymal_hyper: float
+    myymal_super: float
+    myymal_pien: float
+    myymal_muu: float
+    majoit: float
+    asla: float
+    ravint: float
+    tsto: float
+    liiken: float
+    hoito: float
+    kokoon: float
+    opetus: float
+    muut: float
 
 
 class services_transport_km(SQLModel, table=True):
